moved symbols -> win32/symbols
