+++ Modules/Symbols/Include/Win32/Resolver.hpp
