+++ pwn++/pwn.hpp.in
@@ -10,9 +10,11 @@
 #include "Utils.hpp"
 #include "Handle.hpp"
 #include "Context.hpp"
-
-#include "Disassembler.hpp"
 #include "Crypto.hpp"
+
+#ifdef PWN_INCLUDE_DISASSEMBLER
+#include "Disassembler.hpp"
+#endif // PWN_INCLUDE_DISASSEMBLER
 
 #if defined( PWN_BUILD_FOR_WINDOWS)
 #include "Win32/Network.hpp"
@@ -29,8 +31,8 @@
 #include "Win32/Service.hpp"
 #include "Win32/ALPC.hpp"
 #include "Win32/RPC.hpp"
-#include "API.hpp"
-#include "Symbols.hpp"
+#include "Win32/API.hpp"
+#include "Win32/Symbols.hpp"
 
 #include "CTF/Win32/Remote.hpp"
 #include "CTF/Win32/Process.hpp"
